Add Combo & Fix shot mechanism

+++ src/engine/GameSettings.java
@@ -17,7 +17,8 @@ public class GameSettings {
 	/** Frequency of enemy shootings, +/- 30%. */
 	private int shootingFrecuency;
 	/** Difficulty of Game */
-	private double difficulty;
+
+	private int difficulty;
 	/** Check if the stage is boss stage. */
 	private boolean isBossStage;
